Fix db lifecycle for import test classes
